rag_claude/requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0

# External services
anthropic

# Embeddings (HuggingFace)
sentence-transformers>=2.2.2

# Vector Database
langchain-chroma==0.2.6
chromadb==1.1.0

# Utilities
python-dotenv==1.0.0
